FIX query-docs in docCache
