+++ package.json
@@ -41,7 +41,8 @@
     "build": "npm run transpile && browserify dist/lib/browserify.index.js > dist/rxdb.browserify.js && uglifyjs --compress --mangle --output dist/rxdb.browserify.min.js -- dist/rxdb.browserify.js",
     "disc": "npm run transpile && browserify dist/lib/index.js --full-paths | uglifyjs --compress --mangle | discify --open --full-paths",
     "preversion": "npm run lint && npm run test",
-    "dev": "watch 'npm run test:node' src/ test/"
+    "dev": "watch 'npm run test:node' src/ test/",
+    "dev:example": "watch 'npm run transpile:src && echo \"done\"' src/ test/"
   },
   "pre-commit": [
     "lint"
